# swagger/api.yaml
openapi: '3.0.3'
info:
  title: '8 API'
  version: '0.0.1'
paths:
...
  /api/auth/deleteAccount:
    post:
      summary: 'Delete a user'
      operationId: 'deleteAccount'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 'User deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthResponse'
...
components:
  schemas:
...
    UserAuthResponse:
      type: 'object'
      properties:
        token:
          type: 'string'
          example: '1234-1234-1234-1234'
...